Add behaviour pipeline

+++ TreeAndJournal.Api/Exceptions/ExceptionDetails.cs
@@ -1,4 +1,4 @@
 ﻿namespace TreeAndJournal.Api.Exceptions
 {
-    public record ExceptionDetails(int StatusCode, string Type, int Id,string Data);
+    public record ExceptionDetails(int StatusCode, string Type, long Id,string Data);
 }
