+++ TreeAndJournal.Application/Journal/GetItem/JournalDto.cs
@@ -2,8 +2,8 @@
 {
     public class JournalDto
     {
-        public int Id { get; set; }
-        public int EventId { get; set; }
+        public long Id { get; set; }
+        public long EventId { get; set; }
         public string Text { get; set; }
         public DateTime CreatedAt { get; set; }
     }
